fixed issue when buying unavailable token
The bug that was in the button that is in the Details.view.js view has
been eliminated. This button disappears if the token is not available,
and in case of an auction it redirected you to the page where that token
is being auctioned. In the Auction.view.js view, redirect the user in
case the token is not available as an auction or if it is available as a
sale.